requirements.txt
pywebview>=4.0.0
requests>=2.28.0
Pillow>=9.0.0
pystray>=0.19.0

